+++ apps/core/urls.py
@@ -3,14 +3,22 @@ from django.urls import path
 from apps.core import views
 
 urlpatterns = [
-    
     # App pages
     path("home", views.HomeView.as_view(), name="home"),
     path("settings", views.UserSettingsView.as_view(), name="settings"),
     path("admin-panel", views.AdminPanelView.as_view(), name="admin_panel"),
     # Utils
     path("settings/api-key/rotate/", views.rotate_api_key, name="rotate_api_key"),
+    path(
+        "settings/github/starred/import/",
+        views.import_starred_repositories,
+        name="import_starred_repositories",
+    ),
+    path(
+        "settings/github/starred/disable/",
+        views.disable_starred_repository_import,
+        name="disable_starred_repository_import",
+    ),
     path("resend-confirmation/", views.resend_confirmation_email, name="resend_confirmation"),
     path("delete-account/", views.delete_account, name="delete_account"),
-    
 ]